General updates in preparation for 1.1.0. * No Google chagnes * Added support for Metrix (default disabled, opt-in) * Cleaned up Help a little bit. * Hide command menu before loading icons. * Command menu does not flicker after a page load.

+++ install.sh
@@ -1,5 +1,40 @@
 #!/bin/bash
 
+STAGING="staging/"
+METRIX="MetrixLibrary/"
+
+if [ ! -d ${METRIX} ]; then
+    echo "Please download the Metrix libraries"
+    exit
+fi;
+
+echo "**** Cleaning Universe"
+
 rm com.openmobl.app.universe_*
-palm-package --use-v1-format com.openmobl.app.universe
+
+if [ ! -d ${STAGING} ]; then
+    mkdir ${STAGING}
+else
+    rm -rf ${STAGING}
+    mkdir ${STAGING}
+fi;
+
+echo "**** Staging Universe"
+
+cp -R com.openmobl.app.universe/ ${STAGING}
+
+echo "**** Staging Metrix"
+
+cp -R ${METRIX}images ${STAGING}
+cp ${METRIX}app/models/metrix.js ${STAGING}app/models/
+cp ${METRIX}app/models/metrixCore.js ${STAGING}app/models/
+cp ${METRIX}app/models/asyncWrappers.js ${STAGING}app/models/
+cp -R ${METRIX}app/views/metrix ${STAGING}app/views/
+
+echo "**** Packaging Universe"
+
+palm-package ${STAGING}
+
+echo "**** Installing Universe"
+
 palm-install com.openmobl.app.universe_*
